+++ contracts/libraries/EtherAsTokenLibrary.sol
@@ -1,0 +1,6 @@
+//SPDX-License-Identifier: MIT
+pragma solidity >=0.5.0 <0.9.0;
+
+library EtherAsTokenLibrary {
+    address internal constant ETHER_AS_TOKEN = address(0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE);
+}