+++ test/liquidity-accumulator/supply-borrow-accumulator.js
@@ -227,6 +227,21 @@ describe("CompoundV2SBAccumulator#refreshTokenMappings", function () {
         await expect(accumulator.refreshTokenMappings()).to.be.revertedWith("DuplicateMarket");
     });
 
+    it("Reverts if two different CEther tokens are listed", async function () {
+        const cEther1 = await cTokenFactory.deploy(WETH);
+        await cEther1.deployed();
+        const cEther2 = await cTokenFactory.deploy(WETH);
+        await cEther2.deployed();
+
+        await cEther1.stubSetIsCEther(true);
+        await cEther2.stubSetIsCEther(true);
+
+        await poolStub["stubAddMarket(address)"](cEther1.address);
+        await poolStub["stubAddMarket(address)"](cEther2.address);
+
+        await expect(accumulator.refreshTokenMappings()).to.be.revertedWith("DuplicateMarket");
+    });
+
     it("Discovers one market - WETH", async function () {
         const cToken = await cTokenFactory.deploy(WETH);
         await cToken.deployed();
@@ -403,13 +418,26 @@ describe("CompoundV2SBAccumulator#refreshTokenMappings", function () {
         expect(receipt.events.length).to.equal(1);
     });
 
-    it("CTokens without underlying tokens are ignored", async function () {
+    it("CTokens without underlying tokens are treated as CEther", async function () {
         const cToken = await cTokenFactory.deploy(WETH);
         await cToken.deployed();
 
         await cToken.stubSetIsCEther(true);
 
         await poolStub["stubAddMarket(address)"](cToken.address);
+        const refreshTx = await accumulator.refreshTokenMappings();
+        const receipt = await refreshTx.wait();
+
+        expect(refreshTx).to.emit(accumulator, "TokenMappingsRefreshed").withArgs(1, 0);
+        expect(refreshTx).to.emit(accumulator, "CTokenAdded").withArgs(cToken.address);
+        expect(receipt.events.length).to.equal(2);
+    });
+
+    it("Skips markets whose cToken is address(0)", async function () {
+        const cToken = await cTokenFactory.deploy(WETH);
+        await cToken.deployed();
+
+        await poolStub["stubAddMarket(address)"](ethers.constants.AddressZero);
         const refreshTx = await accumulator.refreshTokenMappings();
         const receipt = await refreshTx.wait();
 
@@ -487,8 +515,84 @@ function createDescribeCompoundV2FetchLiquidityTests(typicalSupplyCalculation) {
     };
 }
 
+function createDescribeSpecificVenusIsolatedFetchLiquidityTests() {
+    return function describeVenusIsolatedFetchLiquidityTests(contractName, deployContracts) {
+        describe("Venus isolated special cases", function () {
+            var poolStub;
+            var accumulator;
+            var decimals;
+            var cTokenFactory;
+
+            beforeEach(async function () {
+                const contracts = await deployContracts(WETH, (decimals = DEFAULT_DECIMALS));
+                poolStub = contracts.poolStub;
+                accumulator = contracts.accumulator;
+                cTokenFactory = await ethers.getContractFactory("IonicCTokenStub");
+            });
+
+            const n = 1000;
+
+            it("Fuzz testing with " + n + " iterations", async function () {
+                const token = WETH;
+                const updateData = ethers.utils.hexZeroPad(token, 32);
+
+                const tokenContract = await ethers.getContractAt("IERC20Metadata", token);
+                const tokenDecimals = await tokenContract.decimals();
+
+                for (let i = 0; i < n; i++) {
+                    // Generate a random number of supply, borrow, reserves, and bad debt
+                    const cash = getRandomBigNumber(64);
+                    const borrow = getRandomBigNumber(64);
+                    const reserves = getRandomBigNumber(64);
+                    const badDebt = getRandomBigNumber(64);
+
+                    const totalSupply = cash.add(borrow).add(badDebt).sub(reserves);
+                    const totalBorrow = borrow.add(badDebt);
+
+                    if (totalSupply.lt(0)) {
+                        // Invalid total supply. Retry.
+                        --i;
+
+                        continue;
+                    }
+
+                    const rawCash = ethers.utils.parseUnits(cash.toString(), tokenDecimals - decimals);
+                    const rawBorrow = ethers.utils.parseUnits(borrow.toString(), tokenDecimals - decimals);
+                    const rawReserves = ethers.utils.parseUnits(reserves.toString(), tokenDecimals - decimals);
+                    const rawBadDebt = ethers.utils.parseUnits(badDebt.toString(), tokenDecimals - decimals);
+
+                    // Get the cToken
+                    const cTokenAddress = await poolStub.cTokensByUnderlying(token);
+                    const cToken = await cTokenFactory.attach(cTokenAddress);
+
+                    await cToken.stubSetCash(rawCash);
+                    await cToken.stubSetTotalBorrows(rawBorrow);
+                    await cToken.stubSetTotalReserves(rawReserves);
+                    await cToken.stubSetBadDebt(rawBadDebt);
+
+                    const result = await accumulator.stubFetchLiquidity(updateData);
+
+                    // Allow for a 0.00001% difference
+                    expect(result.tokenLiquidity).to.be.closeTo(totalBorrow, totalBorrow.div(10000000));
+                    expect(result.quoteTokenLiquidity).to.be.closeTo(totalSupply, totalSupply.div(10000000));
+                }
+            });
+        });
+    };
+}
+
 function createDescribeIonicFetchLiquidityTests() {
     return createDescribeCompoundV2FetchLiquidityTests(false);
+}
+
+function createDescribeVenusIsolatedFetchLiquidityTests() {
+    const stdDescribe = createDescribeCompoundV2FetchLiquidityTests(true);
+    const venusIsolatedDescribe = createDescribeSpecificVenusIsolatedFetchLiquidityTests();
+
+    return (contractName, deployContracts) => {
+        stdDescribe(contractName, deployContracts);
+        venusIsolatedDescribe(contractName, deployContracts);
+    };
 }
 
 function describeSBAccumulatorTests(
@@ -798,6 +902,40 @@ async function deployIonicContracts(baseToken, decimals) {
     };
 }
 
+async function deployVenusIsolatedContracts(baseToken, decimals) {
+    const poolStubFactory = await ethers.getContractFactory("IonicStub");
+    const averagingStrategyFactory = await ethers.getContractFactory("ArithmeticAveraging");
+    const accumulatorFactory = await ethers.getContractFactory("VenusIsolatedSBAccumulatorStub");
+    const cTokenFactory = await ethers.getContractFactory("IonicCTokenStub");
+
+    const poolStub = await poolStubFactory.deploy();
+    await poolStub.deployed();
+
+    const averagingStrategy = await averagingStrategyFactory.deploy();
+    await averagingStrategy.deployed();
+
+    const accumulator = await accumulatorFactory.deploy(
+        averagingStrategy.address,
+        poolStub.address,
+        decimals,
+        DEFAULT_UPDATE_THRESHOLD,
+        DEFAULT_UPDATE_DELAY,
+        DEFAULT_HEARTBEAT
+    );
+
+    const cToken = await cTokenFactory.deploy(baseToken);
+    await cToken.deployed();
+
+    await poolStub.stubSetCToken(baseToken, cToken.address);
+
+    await accumulator.refreshTokenMappings();
+
+    return {
+        poolStub: poolStub,
+        accumulator: accumulator,
+    };
+}
+
 async function deployCompoundV2Contracts(baseToken, decimals) {
     const poolStubFactory = await ethers.getContractFactory("IonicStub");
     const averagingStrategyFactory = await ethers.getContractFactory("ArithmeticAveraging");
@@ -976,4 +1114,14 @@ describeSBAccumulatorTests(
     true,
     false,
     createDescribeCompoundV2FetchLiquidityTests(true)
+);
+describeSBAccumulatorTests(
+    "VenusIsolatedSBAccumulator",
+    deployVenusIsolatedContracts,
+    ionicSetSupply,
+    ionicSetBorrow,
+    [_WETH, _USDC],
+    true,
+    false,
+    createDescribeVenusIsolatedFetchLiquidityTests()
 );